+++ Content.Shared/Toggleable/ToggleActionEvent.cs
@@ -1,4 +1,5 @@
 using Content.Shared.Actions;
+using Robust.Shared.Serialization;
 
 namespace Content.Shared.Toggleable;
 
@@ -6,3 +7,13 @@ namespace Content.Shared.Toggleable;
 ///     Generic action-event for toggle-able components.
 /// </summary>
 public sealed class ToggleActionEvent : InstantActionEvent { }
+
+/// <summary>
+///     Generic enum keys for toggle-visualizer appearance data & sprite layers.
+/// </summary>
+[Serializable, NetSerializable]
+public enum ToggleVisuals : byte
+{
+    Toggled,
+    Layer
+}
