+++ Resources/Prototypes/Entities/Objects/Weapons/security.yml
@@ -6,7 +6,10 @@
   components:
   - type: Sprite
     sprite: Objects/Weapons/Melee/stunbaton.rsi
-    state: stunbaton_off
+    netsync: false
+    layers:
+    - state: stunbaton_off
+      map: [ "enum.ToggleVisuals.Layer" ]
   - type: Stunbaton
     energyPerUse: 100
   - type: MeleeWeapon
@@ -33,6 +36,13 @@
     - Belt
   - type: DisarmMalus
     malus: 0.225
+  - type: Appearance
+  - type: GenericVisualizer
+    visuals:
+     enum.ToggleVisuals.Toggled:
+        enum.ToggleVisuals.Layer:
+          True: {state: stunbaton_on}
+          False: {state: stunbaton_off}
 
 - type: entity
   name: flash
